Make tests work on Rails 4.0
The output format of Sprocket/Sass has changed and doesn't include
the partial name anymore. The custom headers include the partial name
and will be output in the CSS unless the output format is :compressed.

This change should not hurt in development and is invisible in
production environments where comments and whitespace will be removed.